refactor(util): consolidate the duplicated mock HTTP server, jitter source, and char-truncation helper
The post-M1 consolidation digest from the phase-1 review (#180), pulled
forward now that the stack is at its final pre-merge state.

Mock server: the ~130-line TCP mock that rss/testutil.rs forked from
open_connector/testutil.rs moves to util/mock_http.rs as the superset
both suites need (byte bodies, no auto-injected header, scripted delay,
mid-transfer truncation, full reason-phrase table, request-body
recording). Both testutils become re-exports plus one flavor
constructor each (ok for gateway envelopes, xml for feeds); every
existing constructor call site and .url access stays as written. The
gateway mock's hardcoded content-type: application/json is gone —
nothing in the client or its tests reads a response content type. The
shared module carries its own harness tests, which also keep the
rss-only response features exercised in test builds without the rss
feature.

Backoff: only the jitter *source* is hoisted (clock_jitter_nanos), per
util::http's own doc that retry loops stay with their callers. The two
spreads (flat 0-100ms vs the fetcher's spec'd ±50%) and the retry
constants remain deliberately local; no timing behavior changes.

Truncation: rss error.rs's truncate becomes the crate-shared
util::text::truncate_chars, re-exported under its old name; the two
inline chars().take() bounds in the Open Connector client and exec now
call it.

The digest's read_body item is deliberately not taken: the shared
skeleton is ~8 lines and the semantics differ at the core (OC prechecks
Content-Length; rss meters the decoded stream only).
