+++ crates/skardi/src/util/mod.rs
@@ -3,3 +3,6 @@
 pub mod http;
 pub mod json;
 pub mod json_pack;
+#[cfg(test)]
+pub(crate) mod mock_http;
+pub mod text;
